requirements.txt
streamlit
openai
python-dotenv
firebase-admin
pyrebase4
google-cloud-storage
